+++ e-Look/src/main/java/com/e_Look/reportMessage/model/ReportMessageDAO_JNDI.java
@@ -234,7 +234,7 @@ public class ReportMessageDAO_JNDI implements ReportMessageDAO_interface {
 	}
 
 	@Override
-	public String getJSON(Integer status) {
+	public String getJSON(Byte status) {
 		//List<ReportMessageVO> list = new ArrayList<ReportMessageVO>();
 		String jsonString;
 		//ReportMessageVO reportMessageVO = null;
